Moved all imports and functionality related to pulling the guesses from Firebase to the Guesses component.
