+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\sirpo\\projects\\babyGamble\\src\\pages\\Home.js":"1","C:\\Users\\sirpo\\projects\\babyGamble\\src\\components\\Firebase\\index.js":"2","C:\\Users\\sirpo\\projects\\babyGamble\\src\\App.js":"3","C:\\Users\\sirpo\\projects\\babyGamble\\src\\components\\Guesses\\Guesses.js":"4"},{"size":796,"mtime":1607284129102,"results":"5","hashOfConfig":"6"},{"size":154,"mtime":1607283683704,"results":"7","hashOfConfig":"6"},{"size":704,"mtime":1607283434432,"results":"8","hashOfConfig":"6"},{"size":541,"mtime":1607283866688,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1ei8dvv",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\sirpo\\projects\\babyGamble\\src\\pages\\Home.js",[],"C:\\Users\\sirpo\\projects\\babyGamble\\src\\components\\Firebase\\index.js",[],"C:\\Users\\sirpo\\projects\\babyGamble\\src\\App.js",[],"C:\\Users\\sirpo\\projects\\babyGamble\\src\\components\\Guesses\\Guesses.js",[]]
+[{"C:\\Users\\sirpo\\projects\\babyGamble\\src\\pages\\Home.js":"1","C:\\Users\\sirpo\\projects\\babyGamble\\src\\components\\Firebase\\index.js":"2","C:\\Users\\sirpo\\projects\\babyGamble\\src\\App.js":"3","C:\\Users\\sirpo\\projects\\babyGamble\\src\\components\\Guesses\\Guesses.js":"4"},{"size":377,"mtime":1607284506664,"results":"5","hashOfConfig":"6"},{"size":154,"mtime":1607283683704,"results":"7","hashOfConfig":"6"},{"size":704,"mtime":1607283434432,"results":"8","hashOfConfig":"6"},{"size":920,"mtime":1607284488567,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1ei8dvv",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\sirpo\\projects\\babyGamble\\src\\pages\\Home.js",[],"C:\\Users\\sirpo\\projects\\babyGamble\\src\\components\\Firebase\\index.js",[],"C:\\Users\\sirpo\\projects\\babyGamble\\src\\App.js",[],"C:\\Users\\sirpo\\projects\\babyGamble\\src\\components\\Guesses\\Guesses.js",[]]
