+++ src/components/Guesses/Guesses.js
@@ -1,8 +1,21 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
+import { firestore, collectIdsAndDocs } from '../Firebase';
 
 import Container from '@material-ui/core/Container';
 
-const Guesses = ({ guesses }) => {
+const Guesses = () => {
+  const [guesses, setGuesses] = useState([]);
+
+  useEffect(() => {
+    getGuesses();
+  }, []);
+
+  async function getGuesses() {
+    const snapshots = await firestore.collection('guesses').get();
+    const userGuesses = snapshots.docs.map(collectIdsAndDocs);
+    setGuesses(userGuesses);
+  }
+
   return (
     <Container>
       {!!guesses &&
